fix(rb-kafka-health): fix RDKafkaErrorCode variant + extract test module
- Remove non-existent `RDKafkaErrorCode::TimedOut` from `is_reqtmout()`.
  The variant does not exist in rdkafka-sys 4.10; `OperationTimedOut`
  covers the local REQTMOUT / ApiVersionRequest timeout scenario.
- Extract `#[cfg(test)] mod tests` from watcher.rs (was 745 lines) into
  watcher_tests.rs via `#[path]`, bringing watcher.rs to 302 lines.
